add send and add mods
